# Refactor crate features for more control over hash-maps usage