--- a/crates/collections/Cargo.toml
+++ b/crates/collections/Cargo.toml
@@ -14,24 +14,44 @@ categories.workspace = true
 exclude.workspace = true
 
 [dependencies]
-hashbrown = { version = "0.14", default-features = false, features = ["ahash", "inline-more"] }
-string-interner = { version = "0.17", default-features = false, features = ["inline-more", "backends"] }
-ahash = { version = "0.8.11", default-features = false }
+hashbrown = { version = "0.14", default-features = false, optional = true, features = ["ahash", "inline-more"] }
+string-interner = { version = "0.17", default-features = false, optional = true, features = ["inline-more", "backends"] }
+ahash = { version = "0.8.11", default-features = false, optional = true }
 
 [features]
-default = ["std", "no-hash-maps"]
+default = ["std"]
 std = ["string-interner/std"]
-# Tells the `wasmi_collections` crate to avoid using hash based maps and sets.
-# 
-# Some embedded environments cannot provide a random source which is required
-# to properly initialize hashmap based data structures for resilience against
-# malious actors that control their inputs.
+
+# Hash collections usage:
 #
-# An example of such an environment is `wasm32-unknown-unknown`.
-no-hash-maps = []
+# - Enable `hash-collections` to make use of hash-based collections in `wasmi_collections`.
+# - Enable `prefer-btree-collections` to still use btree-based collections even when
+#   the `hash-collections` crate feature is enabled.
+#
+# Note:
+#
+# - Not enabling `hash-collections` allows `wasmi_collections` to drop lots of
+#   hash-based dependencies and thus decrease compilation times significantly.
+# - Btree-based collections can be useful for environments without a random source.
+#
+# Which collections will be used:
+#
+# `hash-collections` | `prefer-btree-collections` |      usage
+# ------------------ | -------------------------- | -------------------
+#            false   |                    false   |      btree
+#             true   |                    false   |      hash
+#            false   |                     true   |      btree
+#             true   |                     true   |      btree
+#
+hash-collections = [
+  'dep:hashbrown',
+  'dep:string-interner',
+  'dep:ahash',
+]
+prefer-btree-collections = []
 
 [package.metadata.cargo-udeps.ignore]
 normal = [
-    # The string-interner dependency is always specified even though it is unused when no-hash-maps is enabled.
+    # Needed to suppress weird `udep` warnings. Maybe a `cargo-udep` bug?
     "string-interner"
 ]