This allows to drop some `wasmi_collections` dependencies when unused for improved compile times.

## Usage

Wasmi users who previously used `no-hash-maps` should simply use `prefer-btree-collections` now instead to make sure to always use btree-based collections in their compiled artifacts.

## Summary

- Removes `no-hash-maps` crate feature. Also remove it from `default` feature set.
- Adds `hash-collections` crate feature.
- Adds `prefer-btree-collections` crate feature.

Feature/collections usage matrix:

| `hash-collections` | `prefer-btree-collections` | usage |
|:---:|:---:|:---|
| false | false | btree (default) |
| true | false | hash |
| false | true | btree |
| true | true | btree |

Similar to: https://github.com/bytecodealliance/wasm-tools/pull/1866/files
Will be even more useful once https://github.com/wasmi-labs/wasmi/pull/1141 has been merged.

## Benchmarks

| | Before | After |
|:--:|:--:|:--:|
| Compiled Crates | 22 | 15 |
| Build Time (`dev`) | 5.8s | 5.5s |
| Build Time (`release`) | 7.9s | 7.7s |